requirements.txt
flask==3.0.0
pydantic==2.5.0
openai==1.51.0
python-dotenv==1.0.0
